登录调整，整合重复提交的代码

+++ sources/ci_shan/src/main/java/org/alvin/cishan/sys/auth/adminrole/AdminRoleCond.java
@@ -34,6 +34,6 @@ public class AdminRoleCond extends BaseCondition {
 	private Integer id;// 主键不等于
 	private String name;// 名称
 	private String remark;// 备注
-	private List<Long> ids;// 主键列表
+	private List<Integer> ids;// 主键列表
 
 }
